// backend/src/types/permission-action.ts
export enum PermissionAction {
    READ = 'read',
    CHANGE = 'change',
    MANAGE = 'manage',
    FIND = 'find',
    GET = 'get',
    CREATE = 'create',
    UPDATE = 'update',
    PATCH = 'patch',
    REMOVE = 'remove'
}

export enum PermissionDataAction {
    CHANGE_DATA = 'change.data',
    MANAGE_DATA = 'manage.data',
    CREATE_DATA = 'create.data',
    UPDATE_DATA = 'update.data',
    PATCH_DATA = 'patch.data'
}